ToDo List - styles refactoring(first step)
