+++ todo-list/src/components/AppHeader/AppHeader.js
@@ -1,9 +1,10 @@
 import React from "react";
+import "./AppHeader.scss";
 
 export default function AppHeader() {
   return (
     <div className="header">
-      <h1>TODO List</h1>
+      <h1 className="header-headline">TODO List</h1>
     </div>
   );
 }
