+++ todo-list/src/components/AppHeader/AppHeader.scss
@@ -1,0 +1,7 @@
+.header {
+  margin-bottom: 30px;
+
+  .header-headline {
+    text-align: center;
+  }
+}
