+++ todo-list/src/styles/variables.scss
@@ -1,0 +1,5 @@
+$color-done: lightgreen;
+$color-in-progress: lightskyblue;
+$color-list-border: rgb(243, 241, 241);
+$color-font-main: rgb(87, 87, 87);
+$tranistion-duration: 0.3s;
